Bump to v0.0.2 due to new computational strategy
I was getting a lot of silent dropout when deploying this to a toy
machine.  To debug this, I needed a bit better debugging support.
Going from strace, it seems like the sound device has just stopped
responding, so I refactored the code to accumulate in the callbacks,
then act on that in the main loop.

This changes the way the data gets computed in a meaningful way, so
I'm bumping the version.

We can now identify when we haven't seen any samples, and after a few
times of that, the process can exit and have systemd restart us up.